New design enabling both XYZ units and angle controllers at the same time.
Need code re-factoring

+++ source/abJoystickAxis.cpp
@@ -83,7 +83,7 @@ double JoyStickAxis::getCurrentVelocity()
 	return mLastSetVelocity;
 }
 
-void JoyStickAxis::Move(double newPosition)
+void JoyStickAxis::Move(int newPosition)
 {
 	if(!mMotor || !isEnabled())
     {
